requirements.txt
paramiko
pydantic
pandas
python-dotenv
openpyxl
xlsxwriter
ttkbootstrap
suds